increase the width to a maximum of 1280px
We had to use a wrapper to get the correct width of the left col.
Because fixed width are always relative to the window we increase the
width of the left col to the same values as the container. The warapper
adds then the margin right relative to this container so that the inner
width matches the default width of 3 columns.